example and test generated files changed
