+++ examples/codegen/a/b.rs
@@ -1,4 +1,4 @@
-//! Automatically generated rust module for 'test_required_pb.proto' file
+//! Automatically generated rust module for 'data_types_import.proto' file
 
 #![allow(non_snake_case)]
 #![allow(non_upper_case_globals)]
@@ -7,23 +7,22 @@
 #![allow(clippy)]
 #![cfg_attr(rustfmt, rustfmt_skip)]
 
-pub mod mod_test_required {
 
 use std::io::Write;
 use quick_protobuf::{MessageWrite, BytesReader, Writer, Result};
 use quick_protobuf::sizeofs::*;
 
 #[derive(Debug, Default, PartialEq, Clone)]
-pub struct TestRequired {
-    pub b: bool,
+pub struct ImportedMessage {
+    pub i: Option<bool>,
 }
 
-impl TestRequired {
+impl ImportedMessage {
     pub fn from_reader(r: &mut BytesReader, bytes: &[u8]) -> Result<Self> {
         let mut msg = Self::default();
         while !r.is_eof() {
             match r.next_tag(bytes) {
-                Ok(40) => msg.b = r.read_bool(bytes)?,
+                Ok(8) => msg.i = Some(r.read_bool(bytes)?),
                 Ok(t) => { r.read_unknown(bytes, t)?; }
                 Err(e) => return Err(e),
             }
@@ -32,16 +31,14 @@ impl TestRequired {
     }
 }
 
-impl MessageWrite for TestRequired {
+impl MessageWrite for ImportedMessage {
     fn get_size(&self) -> usize {
         0
-        + 1 + sizeof_varint(*(&self.b) as u64)
+        + self.i.as_ref().map_or(0, |m| 1 + sizeof_varint(*(m) as u64))
     }
 
     fn write_message<W: Write>(&self, w: &mut Writer<W>) -> Result<()> {
-        w.write_with_tag(40, |w| w.write_bool(*&self.b))?;
+        if let Some(ref s) = self.i { w.write_with_tag(8, |w| w.write_bool(*s))?; }
         Ok(())
     }
-}
-
 }
