--- examples/codegen/data_types_import.rs
@@ -1,49 +1,0 @@
-//! Automatically generated rust module for 'data_types_import.proto' file
-
-#![allow(non_snake_case)]
-#![allow(non_upper_case_globals)]
-#![allow(non_camel_case_types)]
-#![allow(unknown_lints)]
-#![allow(clippy)]
-#![cfg_attr(rustfmt, rustfmt_skip)]
-
-pub mod mod_a {
-pub mod mod_b {
-
-use std::io::Write;
-use quick_protobuf::{MessageWrite, BytesReader, Writer, Result};
-use quick_protobuf::sizeofs::*;
-
-#[derive(Debug, Default, PartialEq, Clone)]
-pub struct ImportedMessage {
-    pub i: Option<bool>,
-}
-
-impl ImportedMessage {
-    pub fn from_reader(r: &mut BytesReader, bytes: &[u8]) -> Result<Self> {
-        let mut msg = Self::default();
-        while !r.is_eof() {
-            match r.next_tag(bytes) {
-                Ok(8) => msg.i = Some(r.read_bool(bytes)?),
-                Ok(t) => { r.read_unknown(bytes, t)?; }
-                Err(e) => return Err(e),
-            }
-        }
-        Ok(msg)
-    }
-}
-
-impl MessageWrite for ImportedMessage {
-    fn get_size(&self) -> usize {
-        0
-        + self.i.as_ref().map_or(0, |m| 1 + sizeof_varint(*(m) as u64))
-    }
-
-    fn write_message<W: Write>(&self, w: &mut Writer<W>) -> Result<()> {
-        if let Some(ref s) = self.i { w.write_with_tag(8, |w| w.write_bool(*s))?; }
-        Ok(())
-    }
-}
-
-}
-}
