+++ examples/codegen/data_types_unit.rs
@@ -3,6 +3,10 @@
 #![allow(non_snake_case)]
 #![allow(non_upper_case_globals)]
 #![allow(non_camel_case_types)]
+#![allow(unknown_lints)]
+#![allow(clippy)]
+#![cfg_attr(rustfmt, rustfmt_skip)]
+
 
 use quick_protobuf::{BytesReader, Result, MessageWrite};
