+++ examples/codegen/mod.rs
@@ -1,6 +1,3 @@
-//! Simple module to etract codegen from actual examples
-
-#![allow(unused)]
-pub mod data_types_import;
-pub mod data_types_unit;
+pub mod a;
 pub mod data_types;
+pub mod data_types_unit;
