+++ tests/rust_protobuf/v2/foo/mod.rs
@@ -1,0 +1,2 @@
+pub mod baz;
+pub mod bar;
