+++ tests/rust_protobuf/v2/mod.rs
@@ -39,3 +39,24 @@ mod test_oneof;
 mod test_sync;
 
 mod struct_pb;
+pub mod struct_pb;
+pub mod basic;
+pub mod test_default_values;
+pub mod test_enum_values_pb;
+pub mod test_ident_pb;
+pub mod test_import_nested_imported_pb;
+pub mod test_import_nested_pb;
+pub mod nonunique_1;
+pub mod nonunique_2;
+pub mod test_import_nonunique_pb;
+pub mod foo;
+pub mod test_import_root_imported_pb;
+pub mod test_import_root_pb;
+pub mod test_lite_runtime;
+pub mod test_nonunique_enum_pb;
+pub mod test_oneof_pb;
+pub mod test_required;
+pub mod test_root_pb;
+pub mod test_sanitize_file_name_pb;
+pub mod special;
+pub mod test_text_format_pb;
