+++ tests/rust_protobuf/v2/special.rs
@@ -1,4 +1,4 @@
-//! Automatically generated rust module for 'test_import_nonunique_1_pb.proto' file
+//! Automatically generated rust module for 'test_special~characters file{name}_pb.proto' file
 
 #![allow(non_snake_case)]
 #![allow(non_upper_case_globals)]
@@ -7,20 +7,17 @@
 #![allow(clippy)]
 #![cfg_attr(rustfmt, rustfmt_skip)]
 
-pub mod mod_nonunique_1 {
 
 use quick_protobuf::{BytesReader, Result, MessageWrite};
 
 #[derive(Debug, Default, PartialEq, Clone)]
-pub struct Nonunique { }
+pub struct Outer { }
 
-impl Nonunique {
+impl Outer {
     pub fn from_reader(r: &mut BytesReader, _: &[u8]) -> Result<Self> {
         r.read_to_end();
         Ok(Self::default())
     }
 }
 
-impl MessageWrite for Nonunique { }
-
-}
+impl MessageWrite for Outer { }
