+++ tests/rust_protobuf/v2/test_import_root_pb.rs
@@ -7,11 +7,11 @@
 #![allow(clippy)]
 #![cfg_attr(rustfmt, rustfmt_skip)]
 
+
 use std::io::Write;
 use quick_protobuf::{MessageWrite, BytesReader, Writer, Result};
 use quick_protobuf::sizeofs::*;
-
-use super::test_import_root_imported_pb::*;
+use super::*;
 
 #[derive(Debug, Default, PartialEq, Clone)]
 pub struct ContainsImported {
