+++ tests/rust_protobuf/v2/test_lite_runtime.rs
@@ -1,14 +1,66 @@
-use quick_protobuf::*;
-use rust_protobuf::hex::{encode_hex, decode_hex};
+//! Automatically generated rust module for 'test_lite_runtime_pb.proto' file
 
-use super::test_lite_runtime_pb::mod_test_lite_runtime::*;
+#![allow(non_snake_case)]
+#![allow(non_upper_case_globals)]
+#![allow(non_camel_case_types)]
+#![allow(unknown_lints)]
+#![allow(clippy)]
+#![cfg_attr(rustfmt, rustfmt_skip)]
 
-#[test]
-fn test_lite_runtime() {
-    let mut m = TestLiteRuntime::default();
-    m.v = Some(10);
-    test_serialize_deserialize!("08 0a", &m, TestLiteRuntime);
 
-    // test it doesn't crash
-    format!("{:?}", m);
+use std::io::Write;
+use quick_protobuf::{MessageWrite, BytesReader, Writer, Result};
+use quick_protobuf::sizeofs::*;
+
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+pub enum EnumTestLiteRuntime {
+    ONE = 1,
+    TWO = 2,
+}
+
+impl Default for EnumTestLiteRuntime {
+    fn default() -> Self {
+        EnumTestLiteRuntime::ONE
+    }
+}
+
+impl From<i32> for EnumTestLiteRuntime {
+    fn from(i: i32) -> Self {
+        match i {
+            1 => EnumTestLiteRuntime::ONE,
+            2 => EnumTestLiteRuntime::TWO,
+            _ => Self::default(),
+        }
+    }
+}
+
+#[derive(Debug, Default, PartialEq, Clone)]
+pub struct TestLiteRuntime {
+    pub v: Option<i32>,
+}
+
+impl TestLiteRuntime {
+    pub fn from_reader(r: &mut BytesReader, bytes: &[u8]) -> Result<Self> {
+        let mut msg = Self::default();
+        while !r.is_eof() {
+            match r.next_tag(bytes) {
+                Ok(8) => msg.v = Some(r.read_int32(bytes)?),
+                Ok(t) => { r.read_unknown(bytes, t)?; }
+                Err(e) => return Err(e),
+            }
+        }
+        Ok(msg)
+    }
+}
+
+impl MessageWrite for TestLiteRuntime {
+    fn get_size(&self) -> usize {
+        0
+        + self.v.as_ref().map_or(0, |m| 1 + sizeof_varint(*(m) as u64))
+    }
+
+    fn write_message<W: Write>(&self, w: &mut Writer<W>) -> Result<()> {
+        if let Some(ref s) = self.v { w.write_with_tag(8, |w| w.write_int32(*s))?; }
+        Ok(())
+    }
 }
