+++ tests/rust_protobuf/v2/test_required.rs
@@ -1,20 +1,44 @@
-use quick_protobuf::*;
+//! Automatically generated rust module for 'test_required_pb.proto' file
 
-use super::test_required_pb::mod_test_required::*;
+#![allow(non_snake_case)]
+#![allow(non_upper_case_globals)]
+#![allow(non_camel_case_types)]
+#![allow(unknown_lints)]
+#![allow(clippy)]
+#![cfg_attr(rustfmt, rustfmt_skip)]
 
 
-#[test]
-fn test_write_missing_required() {
-    let mut buf = Vec::new();
-    let mut writer = Writer::new(&mut buf);
-    // every message implements Default and required fields do have a value
-    TestRequired::default().write_message(&mut writer).unwrap();
+use std::io::Write;
+use quick_protobuf::{MessageWrite, BytesReader, Writer, Result};
+use quick_protobuf::sizeofs::*;
+
+#[derive(Debug, Default, PartialEq, Clone)]
+pub struct TestRequired {
+    pub b: bool,
 }
 
-#[test]
-fn test_read_missing_required() {
-    let bytes = &[];
-    let mut reader = BytesReader::from_bytes(bytes);
-    // every message implements Default and required fields do have a value
-    TestRequired::from_reader(&mut reader, bytes).unwrap();
+impl TestRequired {
+    pub fn from_reader(r: &mut BytesReader, bytes: &[u8]) -> Result<Self> {
+        let mut msg = Self::default();
+        while !r.is_eof() {
+            match r.next_tag(bytes) {
+                Ok(40) => msg.b = r.read_bool(bytes)?,
+                Ok(t) => { r.read_unknown(bytes, t)?; }
+                Err(e) => return Err(e),
+            }
+        }
+        Ok(msg)
+    }
+}
+
+impl MessageWrite for TestRequired {
+    fn get_size(&self) -> usize {
+        0
+        + 1 + sizeof_varint(*(&self.b) as u64)
+    }
+
+    fn write_message<W: Write>(&self, w: &mut Writer<W>) -> Result<()> {
+        w.write_with_tag(40, |w| w.write_bool(*&self.b))?;
+        Ok(())
+    }
 }
