+++ tests/rust_protobuf/v2/test_sanitize_file_name_pb.rs
@@ -7,6 +7,7 @@
 #![allow(clippy)]
 #![cfg_attr(rustfmt, rustfmt_skip)]
 
+
 use quick_protobuf::{BytesReader, Result, MessageWrite};
 
 #[derive(Debug, Default, PartialEq, Clone)]
