+++ tests/rust_protobuf/v3/mod.rs
@@ -13,3 +13,7 @@ mod test_ident_pb;
 
 mod test_oneof_pb;
 mod test_oneof;
+pub mod basic;
+pub mod test_ident_pb;
+pub mod test_map_pb;
+pub mod test_oneof_pb;
